+++ src/PEAKUsageSkills/GameAdapters/ConditionSkillAdapter.cs
@@ -4,7 +4,7 @@ namespace PEAKUsageSkills.GameAdapters
 {
     internal static class ConditionSkillAdapter
     {
-        public static bool TryGetResistanceSkill(CharacterAfflictions.STATUSTYPE statusType, out SkillId skillId)
+        public static bool TryGetToleranceSkill(CharacterAfflictions.STATUSTYPE statusType, out SkillId skillId)
         {
             switch (statusType)
             {
@@ -26,30 +26,36 @@ namespace PEAKUsageSkills.GameAdapters
                 case CharacterAfflictions.STATUSTYPE.Hunger:
                     skillId = SkillId.HungerTolerance;
                     return true;
+                case CharacterAfflictions.STATUSTYPE.Curse:
+                    skillId = SkillId.CurseTolerance;
+                    return true;
+                case CharacterAfflictions.STATUSTYPE.Petrify:
+                    skillId = SkillId.PetrificationResistance;
+                    return true;
                 default:
                     skillId = default;
                     return false;
             }
         }
 
-        public static bool TryGetRecoverySkill(CharacterAfflictions.STATUSTYPE statusType, out SkillId skillId)
+        public static bool TryGetRecoveryTolerance(CharacterAfflictions.STATUSTYPE statusType, out SkillId skillId)
         {
             switch (statusType)
             {
                 case CharacterAfflictions.STATUSTYPE.Poison:
-                    skillId = SkillId.PoisonRecovery;
+                    skillId = SkillId.Toxicology;
                     return true;
                 case CharacterAfflictions.STATUSTYPE.Cold:
-                    skillId = SkillId.ColdRecovery;
+                    skillId = SkillId.ColdTolerance;
                     return true;
                 case CharacterAfflictions.STATUSTYPE.Hot:
-                    skillId = SkillId.HeatRecovery;
+                    skillId = SkillId.HeatTolerance;
                     return true;
                 case CharacterAfflictions.STATUSTYPE.Drowsy:
-                    skillId = SkillId.DrowsyRecovery;
+                    skillId = SkillId.DrowsyTolerance;
                     return true;
                 case CharacterAfflictions.STATUSTYPE.Spores:
-                    skillId = SkillId.SporeRecovery;
+                    skillId = SkillId.SporeTolerance;
                     return true;
                 default:
                     skillId = default;